+++ lithops/scripts/cli.py
@@ -73,9 +73,8 @@ def lithops_cli():
 @click.option('--backend', '-b', default=None, help='compute backend')
 @click.option('--storage', '-s', default=None, help='storage backend')
 @click.option('--debug', '-d', is_flag=True, help='debug mode')
-@click.option('--all', '-a', is_flag=True, help='delete all, including master vm in case of standalone')
-@click.option('--force', '-f', is_flag=True, help='force to delete all')
-def clean(config, backend, storage, debug, all, force):
+@click.option('--all', '-a', is_flag=True, help='delete all, including master VM in case of standalone')
+def clean(config, backend, storage, debug, all):
     if config:
         config = load_yaml_config(config)
 
@@ -101,7 +100,7 @@ def clean(config, backend, storage, debug, all, force):
         compute_config = extract_standalone_config(config)
         compute_handler = StandaloneHandler(compute_config)
 
-    compute_handler.clean(delete_master=all, force=force)
+    compute_handler.clean(all=all)
 
     # Clean object storage temp dirs
     storage = internal_storage.storage
@@ -447,7 +446,7 @@ def build(ctx, name, file, config, backend, debug):
     config = default_config(config, config_ow, load_storage_config=False)
 
     if config['lithops']['mode'] != SERVERLESS:
-        raise Exception('"lithops build" command is only valid for serverless backends')
+        raise Exception('"lithops build" command is only available for serverless backends')
 
     compute_config = extract_serverless_config(config)
     compute_handler = ServerlessHandler(compute_config, None)
@@ -480,7 +479,7 @@ def deploy(name, storage, backend, memory, timeout, config, debug):
     config = default_config(config, config_ow)
 
     if config['lithops']['mode'] != SERVERLESS:
-        raise Exception('"lithops runtime deploy" command is only valid for serverless backends')
+        raise Exception('"lithops runtime deploy" command is only available for serverless backends')
 
     storage_config = extract_storage_config(config)
     internal_storage = InternalStorage(storage_config)
@@ -516,7 +515,7 @@ def list_runtimes(config, backend, storage, debug):
     config = default_config(config, config_ow, load_storage_config=False)
 
     if config['lithops']['mode'] != SERVERLESS:
-        raise Exception('"lithops runtime list" command is only valid for serverless backends')
+        raise Exception('"lithops runtime list" command is only available for serverless backends')
 
     compute_config = extract_serverless_config(config)
     compute_handler = ServerlessHandler(compute_config, None)
@@ -561,7 +560,7 @@ def update(name, config, backend, storage, debug):
     config = default_config(config, config_ow)
 
     if config['lithops']['mode'] != SERVERLESS:
-        raise Exception('"lithops runtime update" command is only valid for serverless backends')
+        raise Exception('"lithops runtime update" command is only available for serverless backends')
 
     storage_config = extract_storage_config(config)
     internal_storage = InternalStorage(storage_config)
@@ -570,7 +569,7 @@ def update(name, config, backend, storage, debug):
 
     runtime_info = compute_handler.get_runtime_info()
     runtime_name = runtime_info['runtime_name']
-    runtime_timeout = runtime_info['runtime_memory']
+    runtime_timeout = runtime_info['runtime_timeout']
 
     logger.info(f'Updating runtime: {runtime_name}')
 
@@ -606,7 +605,7 @@ def delete(name, config, memory, version, backend, storage, debug):
     config = default_config(config, config_ow)
 
     if config['lithops']['mode'] != SERVERLESS:
-        raise Exception('"lithops runtime delete" command is only valid for serverless backends')
+        raise Exception('"lithops runtime delete" command is only available for serverless backends')
 
     storage_config = extract_storage_config(config)
     internal_storage = InternalStorage(storage_config)