+++ lithops/standalone/backends/ibm_vpc/config.py
@@ -1,5 +1,6 @@
 #
-# Copyright Cloudlab URV 2021
+# (C) Copyright Cloudlab URV 2020
+# (C) Copyright IBM Corp. 2023
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,30 +17,16 @@
 
 import uuid
 
-MANDATORY_PARAMETERS_1 = ('endpoint',
-                          'vpc_name',
-                          'resource_group_id',
-                          'key_id',
+MANDATORY_PARAMETERS_1 = ('resource_group_id',
                           'iam_api_key')
 
-MANDATORY_PARAMETERS_2 = ('endpoint',
-                          'image_id',
-                          'vpc_id',
-                          'resource_group_id',
-                          'key_id',
-                          'subnet_id',
-                          'security_group_id',
-                          'iam_api_key')
-
-
-MANDATORY_PARAMETERS_3 = ('endpoint',
-                          'instance_id',
-                          'ip_address',
+MANDATORY_PARAMETERS_3 = ('instance_id',
+                          'floating_ip',
                           'iam_api_key')
 
 DEFAULT_CONFIG_KEYS = {
     'master_profile_name': 'cx2-2x4',
-    'profile_name': 'cx2-2x4',
+    'worker_profile_name': 'cx2-2x4',
     'boot_volume_profile': 'general-purpose',
     'ssh_username': 'root',
     'ssh_password': str(uuid.uuid4()),
@@ -47,9 +34,12 @@ DEFAULT_CONFIG_KEYS = {
     'delete_on_dismantle': True,
     'max_workers': 100,
     'worker_processes': 2,
-    'boot_volume_capacity' : 100
+    'boot_volume_capacity': 100
 }
 
+VPC_ENDPOINT = "https://{}.iaas.cloud.ibm.com"
+
+REGIONS = ["jp-tok", "jp-osa", "au-syd", "eu-gb", "eu-de", "us-south", "us-east", "br-sao", "ca-tor"]
 
 def load_config(config_data):
     if 'ibm' in config_data and config_data['ibm'] is not None:
@@ -61,7 +51,7 @@ def load_config(config_data):
 
     if 'exec_mode' in config_data['standalone'] \
        and config_data['standalone']['exec_mode'] in ['create', 'reuse']:
-        params_to_check = MANDATORY_PARAMETERS_2
+        params_to_check = MANDATORY_PARAMETERS_1
     else:
         params_to_check = MANDATORY_PARAMETERS_3
         config_data['ibm_vpc']['max_workers'] = 1
@@ -70,5 +60,19 @@ def load_config(config_data):
         if param not in config_data['ibm_vpc']:
             msg = f"'{param}' is mandatory in 'ibm_vpc' section of the configuration"
             raise Exception(msg)
+
+    if "profile_name" in config_data['ibm_vpc']:
+        config_data['worker_profile_name'] = config_data['profile_name']
+
+    if "region" not in config_data['ibm_vpc'] and "endpoint" not in config_data['ibm_vpc']:
+        msg = "'region' or 'endpoint' parameter is mandatory in 'ibm_vpc' section of the configuration"
+        raise Exception(msg)
+
+    if "region" in config_data['ibm_vpc']:
+        region = config_data['ibm_vpc']['region']
+        if region not in REGIONS:
+            msg = f"'region' conig parameter in 'ibm_vpc' section must be one of {REGIONS}"
+            raise Exception(msg)
+        config_data['ibm_vpc']['endpoint'] = VPC_ENDPOINT.format(region)
 
     config_data['ibm_vpc']['endpoint'] = config_data['ibm_vpc']['endpoint'].replace('/v1', '')